[TASK] Move resource handler to proper folder
